Import of ISHIGAKI/Catalyst-Plugin-Session-Store-Jifty-DBI-0.02 from CPAN.
gitpan-cpan-distribution: Catalyst-Plugin-Session-Store-Jifty-DBI
gitpan-cpan-version:      0.02
gitpan-cpan-path:         ISHIGAKI/Catalyst-Plugin-Session-Store-Jifty-DBI-0.02.tar.gz
gitpan-cpan-author:       ISHIGAKI
gitpan-cpan-maturity:     released

+++ t/lib/TestApp/Controller/DB.pm
@@ -8,7 +8,7 @@ sub setup : Local {
   my ($self, $c) = @_;
 
   my $database = $c->model('DB')->database;
-  if ( -f $database ) {
+  if ( $database && -f $database ) {
     $c->model('DB')->disconnect;
     unlink $database;
   }
@@ -20,7 +20,7 @@ sub teardown : Local {
   my ($self, $c) = @_;
 
   my $database = $c->model('DB')->database;
-  if ( -f $database ) {
+  if ( $database && -f $database ) {
     $c->model('DB')->disconnect;
     unlink $database;
   }
